use Callable type for filter func
