+++ py/sentry_streams/flink/flink_adapter.py
@@ -1,5 +1,4 @@
-from inspect import get_annotations
-from typing import Any, Callable, MutableMapping
+from typing import Any, MutableMapping
 
 from pyflink.common import Types
 from pyflink.common.serialization import SimpleStringSchema
@@ -28,27 +27,6 @@ class FlinkAdapter(StreamAdapter):
     def __init__(self, config: MutableMapping[str, Any], env: StreamExecutionEnvironment) -> None:
         self.environment_config = config
         self.env = env
-
-    def assert_return_type(self, function: Callable[[Any], bool], expected_type: type) -> None:
-        return_type = get_annotations(function)["return"]
-        assert (
-            return_type is expected_type
-        ), f"Filter function {function.__name__} must return a bool, got {return_type} instead"
-
-    def get_function_from_step(self, step: Step) -> Any:
-        assert hasattr(step, "function")
-        fn_path = step.function
-        mod, cls, fn = fn_path.rsplit(".", 2)
-
-        try:
-            module = get_module(mod)
-
-        except ImportError:
-            raise
-
-        imported_cls = getattr(module, cls)
-        imported_fn = getattr(imported_cls, fn)
-        return imported_fn
 
     def source(self, step: Step) -> Any:
         assert hasattr(step, "logical_topic")
@@ -87,17 +65,28 @@ class FlinkAdapter(StreamAdapter):
         return stream.sink_to(sink)
 
     def map(self, step: Step, stream: Any) -> Any:
-        imported_fn = self.get_function_from_step(step)
+        assert hasattr(step, "function")
+        fn_path = step.function
+        mod, cls, fn = fn_path.rsplit(".", 2)
+
+        try:
+            module = get_module(mod)
+
+        except ImportError:
+            raise
+
+        imported_cls = getattr(module, cls)
+        imported_fn = getattr(imported_cls, fn)
 
         # TODO: Ensure output type is configurable like the schema above
         return stream.map(func=lambda msg: imported_fn(msg), output_type=Types.STRING())
 
     def filter(self, step: Step, stream: Any) -> Any:
-        imported_fn = self.get_function_from_step(step)
-        self.assert_return_type(imported_fn, bool)
+        assert hasattr(step, "function")
+        imported_fn = step.function
+        return_type = imported_fn.__annotations__["return"]
+        assert (
+            return_type is bool
+        ), f"Filter function {imported_fn.__name__} must return a bool, got {return_type} instead"
 
-        # filtered = stream.filter(func=lambda msg: imported_fn(msg))
-        # assert type(filtered) is bool, (
-        #     f"Filter function must return a bool, got {type(filtered)} instead"
-        # )
         return stream.filter(func=lambda msg: imported_fn(msg))
