+++ py/tests/test_pipeline.py
@@ -8,6 +8,7 @@ from sentry_streams.adapters.stream_adapter import RuntimeTranslator, StreamAdap
 from sentry_streams.flink.flink_adapter import FlinkAdapter
 from sentry_streams.pipeline import Filter, KafkaSink, KafkaSource, Map, Pipeline
 from sentry_streams.runner import iterate_edges
+from sentry_streams.user_functions.sample_filter import EventsPiplineFilterFunctions
 
 
 @pytest.fixture(autouse=True)
@@ -164,7 +165,7 @@ def basic_filter() -> tuple[Pipeline, MutableMapping[str, list[dict[str, Any]]]]
         name="myfilter",
         ctx=pipeline,
         inputs=[source],
-        function="sentry_streams.sample_functions.EventsPipelineFunctions.simple_filter",
+        function=EventsPiplineFilterFunctions.simple_filter,
     )
 
     _ = KafkaSink(
@@ -252,36 +253,9 @@ def bad_import_map() -> Pipeline:
     return pipeline
 
 
-def bad_import_filter() -> Pipeline:
-    pipeline = Pipeline()
-
-    source = KafkaSource(
-        name="myinput",
-        ctx=pipeline,
-        logical_topic="logical-events",
-    )
-
-    map = Filter(
-        name="myfilter",
-        ctx=pipeline,
-        inputs=[source],
-        function="sentry_streams.unknown_module.EventsPipelineFunctions.simple_filter",
-    )
-
-    _ = KafkaSink(
-        name="kafkasink",
-        ctx=pipeline,
-        inputs=[map],
-        logical_topic="transformed-events",
-    )
-
-    return pipeline
-
-
-@pytest.mark.parametrize("pipeline", [bad_import_map(), bad_import_filter()])
 def test_import(
     setup_basic_flink_env: tuple[StreamExecutionEnvironment, RuntimeTranslator],
-    pipeline: Pipeline,
+    pipeline: Pipeline = bad_import_map(),
 ) -> None:
     _, translator = setup_basic_flink_env
 
@@ -298,27 +272,26 @@ def bad_type_filter() -> Pipeline:
         logical_topic="logical-events",
     )
 
-    map = Filter(
+    filter = Filter(
         name="myfilter",
         ctx=pipeline,
         inputs=[source],
-        function="sentry_streams.sample_functions.EventsPipelineFunctions.wrong_type_filter",
+        function=EventsPiplineFilterFunctions.wrong_type_filter,
     )
 
     _ = KafkaSink(
         name="kafkasink",
         ctx=pipeline,
-        inputs=[map],
+        inputs=[filter],
         logical_topic="transformed-events",
     )
 
     return pipeline
 
 
-@pytest.mark.parametrize("pipeline", [bad_type_filter()])
 def test_typing(
     setup_basic_flink_env: tuple[StreamExecutionEnvironment, RuntimeTranslator],
-    pipeline: Pipeline,
+    pipeline: Pipeline = bad_type_filter(),
 ) -> None:
     _, translator = setup_basic_flink_env
